Starting layers for ab structure. See full commit.
There was an error trying to make the layers with the current unit cell:

-5.493523078    -0.000012783    9.013134903
-5.490765917     0.000006198    7.031903998
-2.747804419     0.000006198    6.046251835
-2.744858933    -0.000005757    3.009985643
-0.002169046     0.000003422    2.122751619
 0.000000000     0.000000000    0.000000000

because rlayer.sh was not working with x negative components.

To solve temporary this issue I made a test in the aa structure for which rlayer.sh
worked properly. I moved all the structure in the x direction making a sum of the
smallest x coordinate in another folder. The result was exactly the same.

So i did the same kind of displacement in the ab structure making a siplacement of
5.493523078 in the x direction for all the rows havin the followin unit cell:

0.000000000    -0.000012783    9.013134903
0.002757161     0.000006198    7.031903998
2.745718659     0.000006198    6.046251835
2.748664145    -0.000005757    3.009985643
5.491354032     0.000003422    2.122751619
5.493523078     0.000000000    0.000000000

Then i ran rlayer.sh and it worked. Later i ensured making the plot that the
layers was created properly. They was!

Finally i copied all the necesary files to run the layer-by-layer matrix elements
and responses in the folder with the original unit cell.