Cheking kk vs sm vs v + 0.00001 & kpts for UP
Checking again kpts convergence in kk files for the up structure; the
doesn't mathc in the sm files.

The convergence was checked again using 12802 and 11252 kpts. Is Ok!

The value of the velocity is significatively higher using kk than sm
files and the gap matches according to the previous paper when the kk
files was used.

Also, when the V^ab is calculated I used a the value +0.00001
in the denominator to avoid divergences. I made a calculation using now
+0.0001 and the values are the same.

The next step is make again all the plots using now the kk files.